Bug 794940: Visible white line in Nexus 7. [r=mfinkle] [needs-clobber]
UltraBlame original commit: afa64181c0dce00ea2309b71390f01ab8bff5b6a

+++ mobile/android/base/TabsButton.java
@@ -59,7 +59,7 @@ public class TabsButton extends ShapedButton {
 
         int width = getMeasuredWidth();
         int height = getMeasuredHeight();
-        float curve = height * 1.125f;
+        int curve = (int) (height * 1.125f);
 
         
         float left;
